# TEST/PERF: Increase thresholds for test_ucp_wait_mem